requirements.txt
appnope
asttokens
certifi==2024.8.30
charset-normalizer==3.4.0
comm
contourpy
cycler==0.12.1
debugpy
decorator
exceptiongroup
executing
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
idna==3.10
importlib_metadata
ipykernel
ipythonpyt
jedi
Jinja2==3.1.4
jupyter_client 
jupyter_core
kiwisolver==1.4.7
MarkupSafe
matplotlib
matplotlib-inline
mpmath==1.3.0
nest_asyncio
networkx
numpy
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging
pandas
parso
pexpect
pickleshare 
pillow
platformdirs
prompt_toolkit
psutil
ptyprocess
pure_eval
py-cpuinfo==9.0.0
Pygments
pyparsing
python-dateutil
pytz==2024.2
PyYAML==6.0.2
pyzmq
requests==2.32.3
scipy
seaborn==0.13.2
six 
stack-data 
sympy==1.13.1
torch
torchvision
tornado
tqdm==4.67.0
traitlets
typing_extensions
tzdata==2024.2
ultralytics
ultralytics-thop==2.0.11
urllib3==2.2.3
wcwidth
zipp
